+++ app/src/main/java/ua/com/radiokot/photoprism/features/viewer/di/MediaViewerFeatureModule.kt
@@ -22,7 +22,6 @@ import ua.com.radiokot.photoprism.env.data.model.EnvSession
 import ua.com.radiokot.photoprism.features.gallery.di.galleryFeatureModule
 import ua.com.radiokot.photoprism.features.viewer.logic.BackgroundMediaFileDownloadManager
 import ua.com.radiokot.photoprism.features.viewer.logic.DefaultVideoPlayerFactory
-import ua.com.radiokot.photoprism.features.viewer.logic.DeleteGalleryMediaUseCase
 import ua.com.radiokot.photoprism.features.viewer.logic.SetGalleryMediaFavoriteUseCase
 import ua.com.radiokot.photoprism.features.viewer.logic.ThreadPoolBackgroundMediaFileDownloadManager
 import ua.com.radiokot.photoprism.features.viewer.logic.VideoPlayerFactory
@@ -93,6 +92,5 @@ val mediaViewerFeatureModule = module {
         viewModelOf(::VideoPlayerCacheViewModel)
 
         scopedOf(SetGalleryMediaFavoriteUseCase::Factory)
-        scopedOf(DeleteGalleryMediaUseCase::Factory)
     }
 }