+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - Ability to archive or delete a single gallery item from the media viewer
-- Ability to archive gallery items selected with a long press
+- Ability to archive or delete gallery items selected with a long press
 
 ### Fixed
 - Not canceling the download on closing the dialog if the screen was rotated